MapperOWM has been completed. Analyzer layer in progress

+++ src/main/java/com/jean/analyzers/weather/WeatherAnalyzer.java
@@ -3,11 +3,11 @@ package com.jean.analyzers.weather;
 
 public interface WeatherAnalyzer {
 
-	double stabilityChecker(float[] temperatureParams);
+	double stabilityChecker(double[] temperatureParams);
 
 	double windChecker(double temperature, double degrees, double speed);
 
-	double pressureChecker(float[] pressureParams);
+	double pressureChecker(double[] pressureParams);
